Revert "Camera: Add pinch to zoom"
If well, pinch to zoom is really awesome and that stuff
camera already has zoom control on the right side and it's better to have
AOSP implementation slide to gallery on the entire screen that just on the
buttons. This may also break swipe to gallery on LDPI devices
(and even MDPI ?), in wich case buttons area is really small.

patch set 2: fixed commit msg

This reverts commit ef0f9d8f0c3929665925a1ce19ec692a72fa7d99

Change-Id: Ieda8f4784b26a398f626927897a57f1d7b644ffe